Move data:signingrecord to ethereum:signingrecord
