+++ data/dataexchange/build.gradle
@@ -3,7 +3,7 @@ dependencies {
     api 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
     implementation project(':infrastructure:io')
     implementation project(':data:serializer')
-    implementation project(':data:signingrecord')
+    implementation project(':ethereum:signingrecord')
     implementation project(':ethereum:spec')
     implementation project(':infrastructure:logging')
     implementation project(':infrastructure:yaml')
