+++ data/dataexchange/src/main/java/tech/pegasys/teku/data/SlashingProtectionExporter.java
@@ -33,10 +33,10 @@ import org.apache.tuweni.bytes.Bytes48;
 import tech.pegasys.teku.api.schema.BLSPubKey;
 import tech.pegasys.teku.api.schema.PublicKeyException;
 import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.data.signingrecord.ValidatorSigningRecord;
 import tech.pegasys.teku.data.slashinginterchange.Metadata;
 import tech.pegasys.teku.data.slashinginterchange.SigningHistory;
 import tech.pegasys.teku.data.slashinginterchange.SlashingProtectionInterchangeFormat;
+import tech.pegasys.teku.ethereum.signingrecord.ValidatorSigningRecord;
 import tech.pegasys.teku.infrastructure.io.SyncDataAccessor;
 import tech.pegasys.teku.provider.JsonProvider;
