+++ ethereum/signingrecord/src/main/java/tech/pegasys/teku/ethereum/signingrecord/ValidatorSigningRecord.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.data.signingrecord;
+package tech.pegasys.teku.ethereum.signingrecord;
 
 import com.google.common.base.MoreObjects;
 import java.util.Objects;
