+++ ethereum/signingrecord/src/main/java/tech/pegasys/teku/ethereum/signingrecord/ValidatorSigningRecordSerialization.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.data.signingrecord;
+package tech.pegasys.teku.ethereum.signingrecord;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
