+++ ethereum/signingrecord/src/test/java/tech/pegasys/teku/ethereum/signingrecord/ValidatorSigningRecordTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.data.signingrecord;
+package tech.pegasys.teku.ethereum.signingrecord;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
