+++ settings.gradle
@@ -18,7 +18,7 @@ include 'data:provider'
 include 'data:beaconrestapi'
 include 'data:publisher'
 include 'data:serializer'
-include 'data:signingrecord'
+include 'ethereum:signingrecord'
 include 'fuzz'
 include 'infrastructure'
 include 'infrastructure:async'
