+++ teku/build.gradle
@@ -2,9 +2,9 @@ dependencies {
   implementation project(':beacon:validator')
   implementation project(':data')
   implementation project(':data:serializer')
-  implementation project(':data:signingrecord')
   implementation project(':data:publisher')
   implementation project(':data:beaconrestapi')
+  implementation project(':ethereum:signingrecord')
   implementation project(':infrastructure:bls')
   implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:metrics')
